--- a/src/GeoTimeZone/TimeZoneResult.cs
+++ b/src/GeoTimeZone/TimeZoneResult.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 
@@ -9,15 +10,16 @@ namespace GeoTimeZone
     /// </summary>
     public class TimeZoneResult
     {
-        internal TimeZoneResult(params string[] timeZones)
+        internal TimeZoneResult(string[] timeZones)
         {
-            if (timeZones.Length == 0)
-            {
-                throw new ArgumentException("There must be at least one value provided.", nameof(timeZones));
-            }
-
             this.Result = timeZones[0];
             this.AlternativeResults = new ReadOnlyCollection<string>(timeZones.Skip(1).ToList());
+        }
+
+        internal TimeZoneResult(string timeZone)
+        {
+            this.Result = timeZone;
+            this.AlternativeResults = new ReadOnlyCollection<string>(new List<string>());
         }
 
         /// <summary>